moved tkf indel stuff to separate file
